cd2023-proj-107849-108073/requirements.txt
ffmpeg-python==0.2.0
torch>=1.8.1
torchaudio>=0.8.1
demucs>=4.0.0
fastapi==0.96.0
python-multipart==0.0.6 
uvicorn==0.22.0
celery==5.3.0
pydub==0.25.1
tinytag==1.9.0